+++ docs/CHANGELOG.rst
@@ -10,6 +10,10 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+Fixed
+-----
+- Correctly relabel SELinux contexts on user/group files
+
 
 ==================
 3.0.0 - 2017-09-13